4971 Set the SelectedQueue when checkbox or deletebutton clicked

+++ Dev/Dev2.Studio/Triggers/QueueEvents/QueueEventsViewModel.cs
@@ -199,8 +199,12 @@ namespace Dev2.Triggers.QueueEvents
         {
             try
             {
-                var triggerQueue = AsTriggerQueueModel();
-                _resourceRepository.DeleteQueue(triggerQueue);
+                if (!SelectedQueue.IsNewQueue)
+                {
+                    var triggerQueue = AsTriggerQueueModel();
+                    _resourceRepository.DeleteQueue(triggerQueue);
+                }
+                
                 Queues.Remove(SelectedQueue);
                 SelectedQueue = null;
                 IsDirty = false;
